refactor(protected-area): split get/change services
