+++ api/apps/api/src/modules/scenarios/protected-area/index.ts
@@ -1,9 +1,12 @@
+export {
+  invalidProtectedAreaId,
+  ChangeProtectedAreasError,
+} from './selection/selection-update.service';
+
 export { ProtectedAreaModule } from './protected-area.module';
 export {
   ProtectedAreaService,
-  ChangeProtectedAreasError,
   submissionFailed,
-  invalidProtectedAreaId,
 } from './protected-area.service';
 
 export { ProtectedAreaKind } from './protected-area.kind';
